# openapi.yaml
openapi: 3.0.3
info:
  title: IRIA Ecosystem Platform API (Hackathon MVP)
  version: 1.0.0
  description: |
    API contract for the Azerbaijan Startup Ecosystem Platform (IRIA Hackathon MVP).

    Platform goals:
    - Startup registry and ecosystem stakeholder management
    - AI-assisted startup analysis (SWOT, success score, market insights)
    - Similar startup discovery
    - Program application workflows
    - Investor interest tracking
    - KPI dashboards for IRIA/Government
    - Role-based access control and audit visibility
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.iria-ecosystem.az/api/v1
    description: Production (planned)
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    patch:
      tags: [Users]
      summary: Update current user profile
      description: Update generic user/stakeholder profile fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
            example:
              full_name: "Aysel Mammadova"
              organization: "NovaTech"
              bio: "Founder building climate-tech solutions."
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: |
        Returns stakeholder profile (role-aware). Access can be restricted by RBAC.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /startups:
    get:
      tags: [Startups]
      summary: List startups
      parameters:
        - in: query
          name: sector
          schema:
            type: string
        - in: query
          name: stage
          schema:
            $ref: "#/components/schemas/StartupStage"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: min_score
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 100
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Startup list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupListResponse"

    post:
      tags: [Startups]
      summary: Create startup profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartupCreateRequest"
      responses:
        "201":
          description: Startup profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /startups/{startup_id}:
    get:
      tags: [Startups]
      summary: Get startup profile by ID
      parameters:
        - $ref: "#/components/parameters/StartupId"
      responses:
        "200":
          description: Startup profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupProfile"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Startups]
      summary: Update startup profile
      parameters:
        - $ref: "#/components/parameters/StartupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartupUpdateRequest"
      responses:
        "200":
          description: Startup profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /startups/{startup_id}/analysis:
    post:
      tags: [AI Analysis]
      summary: Run AI-assisted startup analysis
      parameters:
        - $ref: "#/components/parameters/StartupId"
      responses:
        "200":
          description: Analysis generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupAnalysis"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    get:
      tags: [AI Analysis]
      summary: Get latest analysis for startup
      parameters:
        - $ref: "#/components/parameters/StartupId"
      responses:
        "200":
          description: Latest analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupAnalysis"
        "404":
          $ref: "#/components/responses/NotFound"
  /startups/{startup_id}/similar:
    get:
      tags: [Similarity]
      summary: Get similar startups
      parameters:
        - $ref: "#/components/parameters/StartupId"
        - in: query
          name: limit
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: Similar startups list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarStartupsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /startups/{startup_id}/interest:
    post:
      tags: [Investor]
      summary: Mark investor interest in a startup
      parameters:
        - $ref: "#/components/parameters/StartupId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestorInterestCreateRequest"
      responses:
        "201":
          description: Interest recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvestorInterest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Already marked as interested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Investor]
      summary: Remove investor interest
      parameters:
        - $ref: "#/components/parameters/StartupId"
      responses:
        "204":
          description: Interest removed
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      in: path
      name: user_id
      required: true
      schema:
        type: string
      example: "usr_01JXYZABC123"
    StartupId:
      in: path
      name: startup_id
      required: true
      schema:
        type: string
      example: "st_01JXYZABC123"
...
  responses:
    BadRequest:
      description: Invalid request payload or query params
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UserRole:
      type: string
      enum: [startup, investor, mentor, iria_admin]
    StartupStage:
      type: string
      enum: [idea, pre_seed, seed, series_a, growth]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
          required: [code, message]
      required: [error]
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
      required: [page, page_size, total]
...
    User:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        organization:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, full_name, email, role, created_at, updated_at]
    UserUpdateRequest:
      type: object
      properties:
        full_name:
          type: string
        organization:
          type: string
        bio:
          type: string
    StartupTraction:
      type: object
      properties:
        users:
          type: integer
        monthly_revenue_azn:
          type: number
          format: float
        pilots:
          type: integer
        partnerships:
          type: integer
    StartupFounder:
      type: object
      properties:
        user_id:
          type: string
        full_name:
          type: string
        title:
          type: string
      required: [full_name]
    StartupCreateRequest:
      type: object
      properties:
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        sector:
          type: string
        stage:
          $ref: "#/components/schemas/StartupStage"
        website_url:
          type: string
          format: uri
        founded_year:
          type: integer
        city:
          type: string
        team_size:
          type: integer
        founders:
          type: array
          items:
            $ref: "#/components/schemas/StartupFounder"
        traction:
          $ref: "#/components/schemas/StartupTraction"
      required: [name, description, sector, stage]
    StartupUpdateRequest:
      type: object
      properties:
        tagline:
          type: string
        description:
          type: string
        sector:
          type: string
        stage:
          $ref: "#/components/schemas/StartupStage"
        website_url:
          type: string
          format: uri
        city:
          type: string
        team_size:
          type: integer
        traction:
          $ref: "#/components/schemas/StartupTraction"
    StartupProfile:
      type: object
      properties:
        id:
          type: string
        startup_id_code:
          type: string
        owner_user_id:
          type: string
        name:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
        sector:
          type: string
        stage:
          $ref: "#/components/schemas/StartupStage"
        website_url:
          type: string
          nullable: true
        founded_year:
          type: integer
          nullable: true
        city:
          type: string
          nullable: true
        team_size:
          type: integer
          nullable: true
        founders:
          type: array
          items:
            $ref: "#/components/schemas/StartupFounder"
        traction:
          $ref: "#/components/schemas/StartupTraction"
        latest_success_score:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, startup_id_code, name, description, sector, stage, created_at, updated_at]
    StartupListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StartupProfile"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [items, meta]
    SWOT:
      type: object
      properties:
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        opportunities:
          type: array
          items:
            type: string
        threats:
          type: array
          items:
            type: string
      required: [strengths, weaknesses, opportunities, threats]
    ScoreFactor:
      type: object
      properties:
        factor:
          type: string
        weight:
          type: number
          format: float
        contribution:
          type: number
          format: float
        explanation:
          type: string
      required: [factor, weight, contribution]
    StartupAnalysis:
      type: object
      properties:
        id:
          type: string
        startup_id:
          type: string
        success_score:
          type: number
          format: float
        score_band:
          type: string
          enum: [low, medium, high]
        swot:
          $ref: "#/components/schemas/SWOT"
        market_snapshot:
          type: string
        positioning_insight:
          type: string
        explainability:
          type: array
          items:
            $ref: "#/components/schemas/ScoreFactor"
        generated_at:
          type: string
          format: date-time
      required: [id, startup_id, success_score, score_band, swot, explainability, generated_at]
    SimilarStartupItem:
      type: object
      properties:
        startup_id:
          type: string
        startup_name:
          type: string
        similarity_score:
          type: number
          format: float
        reasons:
          type: array
          items:
            type: string
      required: [startup_id, startup_name, similarity_score, reasons]
    SimilarStartupsResponse:
      type: object
      properties:
        startup_id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SimilarStartupItem"
      required: [startup_id, items]
...
    InvestorInterestCreateRequest:
      type: object
      properties:
        note:
          type: string
          maxLength: 500
    InvestorInterest:
      type: object
      properties:
        id:
          type: string
        investor_user_id:
          type: string
        startup_id:
          type: string
        note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, investor_user_id, startup_id, created_at]
...